'수정'

+++ src/main/java/kr/co/bestiansoft/ebillservicekg/asset/amsImg/vo/AmsImgVo.java
@@ -13,6 +13,6 @@ public class AmsImgVo extends ComDefaultVO {
     private String fileNm;
     private String orgnlFileNm;
     private String fileExtn;
-    private int fileSz;
+    private Long fileSz;
 
 }
